add VitalsPageStore
